+++ src/containers/AgentStateMonitoring/layout.js
@@ -6,6 +6,7 @@ import { messageSubscribe, messageUnsubscribe } from './windowMessageObservable'
 import SidePanel from '../../containers/SidePanel';
 import AgentStateMonitoringBulkActionsForm from './Bulk';
 import AgentStateMonitoringTable from './Table';
+import { sdkCall } from '../../utils/sdk';
 
 const Wrapper = styled.div`
   display: grid;
@@ -35,6 +36,12 @@ export default class AgentStateMonitoring extends Component {
     this.props.fetchData('skills', this.props.tableType);
     this.props.fetchData('reasonLists', this.props.tableType);
     this.props.startAgentStateMonitoring();
+  }
+
+  componentDidMount() {
+    if (this.props.insideIframe) {
+      sdkCall({ module: 'removeDirtyFormIdFromSessionStorage' });
+    }
   }
 
   componentWillUnmount() {
@@ -85,5 +92,6 @@ AgentStateMonitoring.propTypes = {
   slidingWidth: PropTypes.number,
   id: PropTypes.string,
   className: PropTypes.string,
-  bulkSelectedTotal: PropTypes.object
+  bulkSelectedTotal: PropTypes.object,
+  insideIframe: PropTypes.bool
 };