+++ src/containers/FlowDebugLogs/layout.js
@@ -3,6 +3,7 @@ import styled, { css } from 'styled-components';
 import { PageHeader, SearchIconSVG, WordWrapIconSVG, CopyIconSVG, Toast } from 'cx-ui-components';
 import ReactJson from 'react-json-view';
 import PropTypes from 'prop-types';
+import { sdkCall } from '../../utils/sdk';
 import { copyToClipboard, downloadFile } from 'serenova-js-utils/browser';
 
 const dataViewStyle = css`
@@ -108,6 +109,12 @@ export default class FlowDebugLogs extends Component {
       prettyView: false,
       wordwrap: true
     };
+  }
+
+  componentDidMount() {
+    if (this.props.insideIframe) {
+      sdkCall({ module: 'removeDirtyFormIdFromSessionStorage' });
+    }
   }
 
   handleValueChange = event => {
@@ -221,5 +228,6 @@ FlowDebugLogs.propTypes = {
   fetchReportingEvents: PropTypes.func,
   reportingEvents: PropTypes.object,
   reportingEventsString: PropTypes.string,
-  reportingEventsRawData: PropTypes.any
+  reportingEventsRawData: PropTypes.any,
+  insideIframe: PropTypes.bool
 };