+++ src/containers/SupervisorToolbar/layout.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { messageSubscribe, messageUnsubscribe } from './windowMessageObservable.js';
 import { MutedIconSVG, UnMutedIconSVG, HangUpIconSVG } from 'cx-ui-components';
+import { sdkCall } from '../../utils/sdk';
 
 const HangUpIconSVGWrapper = styled(HangUpIconSVG)`
   margin-left: 10px;
@@ -18,6 +19,12 @@ export default class SupervisorToolbar extends Component {
   componentWillMount() {
     this.props.startSupervisorToolbarSubscriptions();
     messageSubscribe();
+  }
+
+  componentDidMount() {
+    if (this.props.insideIframe) {
+      sdkCall({ module: 'removeDirtyFormIdFromSessionStorage' });
+    }
   }
 
   componentWillUnmount() {
@@ -58,5 +65,6 @@ SupervisorToolbar.propTypes = {
   twilioEnabled: PropTypes.bool.isRequired,
   monitoringStatus: PropTypes.string.isRequired,
   interactionId: PropTypes.string.isRequired,
-  userHasBargeAllCallsPermission: PropTypes.bool.isRequired
+  userHasBargeAllCallsPermission: PropTypes.bool.isRequired,
+  insideIframe: PropTypes.bool
 };