+++ src/redux/modules/form/selectors.js
@@ -8,7 +8,8 @@ import {
   getCurrentEntity,
   getSelectedEntity,
   getSelectedEntityId,
-  getSelectedEntityFormId
+  getSelectedEntityFormId,
+  getSelectedSubEntityFormId
 } from '../entities/selectors';
 import { onFormSubmit, onSubEntityFormSubmit } from '../entities';
 
@@ -25,6 +26,7 @@ export const getCurrentFormInitialValues = state =>
 export const isFormInvalid = state => isInvalid(getSelectedEntityFormId(state))(state);
 export const isFormDirty = state => isDirty(getSelectedEntityFormId(state))(state);
 export const isFormPristine = state => isPristine(getSelectedEntityFormId(state))(state);
+export const isSubEntityFormDirty = state => isDirty(getSelectedSubEntityFormId(state))(state);
 
 export const selectFormInitialValues = state => {
   const entityName = getCurrentEntity(state);