+++ CHANGELOG.md
@@ -28,6 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dispositions
 - Disposition Lists
 
+## [0.70.4] - 2019-12-12
+## Fixed
+- CXV1-20513 - Display alert in config-ui while navigating between config2 pages when the form is dirty.
+- CXV1-20971 - Display confirmation modal in config2 while navigating between pages when the form is dirty.
+
 ## [0.70.3] - 2019-12-11
 ## Fixed
 - CXV1-20921 - Wrong error message when attempting to add an unshared disposition to a shared disposition list.